RE-21: who is "Fixiki"?

+++ internal/profile-service/entity/profile/profileResponse.go
@@ -1,10 +1,10 @@
 package profile
 
 type ProfileResponse struct {
-	ID          int    `json: "id"`
-	Username    string `json:"username" validate:"required,min=5,max=50"`
-	Email       string `json:"email" validate:"required,email"`
-	Description string `json:"description" validate:"min=0,max=200"`
-	Balance     int    `json:"balance" validate:"gte=0"`
-	Role        int    `json:"role" validate:"required,gte=1,lte=2"`
+	ID          int     `json: "id"`
+	Username    string  `json:"username" validate:"required,min=5,max=50"`
+	Email       string  `json:"email" validate:"required,email"`
+	Description string  `json:"description" validate:"min=0,max=200"`
+	Balance     Decimal `json:"balance" validate:"gte=0"`
+	Role        int     `json:"role" validate:"required,gte=1,lte=2"`
 }
